requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.0
qdrant-client==1.6.9
openai==1.3.7
pypdf==3.17.1
python-docx==1.1.0
python-dotenv==1.0.0
requests==2.31.0
pydantic-settings==2.1.0
sentence-transformers==2.2.2
gunicorn==21.2.0